new error type, moved errors into types, moved AnyFunction into type
