+++ src/types/function.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type AnyFunction = ((...args: any[]) => any)
