+++ src/types/index.ts
@@ -1,4 +1,7 @@
+export * from './function'
 export * from './mapper'
 export * from './profile'
+export * from './profileNotFoundError'
+export * from './profileTypeError'
 export * from './promise'
 export * from './register'
