+++ src/types/profileNotFoundError.ts
