+++ src/types/profileTypeError.ts
@@ -1,0 +1,5 @@
+export class ProfileTypeError extends Error {
+  public constructor() {
+    super('Every value provided to loadProfiles must implement Profile')
+  }
+}
